fix:action_failed

+++ srcs/parse/calc_map_size.c
@@ -6,7 +6,7 @@
 /*   By: tshimizu <tshimizu@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/25 17:27:20 by tshimizu          #+#    #+#             */
-/*   Updated: 2026/04/25 18:17:43 by tshimizu         ###   ########.fr       */
+/*   Updated: 2026/04/25 18:26:10 by tshimizu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,9 +23,9 @@ void	calc_map_size(t_game *game)
 	width = 0;
 	while (game->map->map_data[height])
 	{
-		len = ft_strlen(game->map->map_data[height]);
+		len = (int)ft_strlen(game->map->map_data[height]);
 		if (width < len)
-			width = ft_strlen(current_line);
+			width = (int)ft_strlen(current_line);
 		height++;
 	}
 	game->map->height = height;
